db-migration/mysql/requirements.txt
boto3
botocore
fabric3
paramiko
progressbar
pwinput
python-dateutil
requests
termcolor
tqdm